First increment of the petition-scanner redesign — the mockup direction
the team approved. Mobile-first, in the Sunflower brand.

## What's in

- **Global scan FAB** (`components/ScanFab.tsx`) — a camera button fixed
bottom-right on every authenticated screen, linking to
`/petition/capture`. Hidden for signed-out users and on the
petition/auth routes. **Replaces the "Petitions" header nav link**
(removed). i18n + `aria-label` + focus ring + 56px target. Unit-tested
(7 cases: renders on normal routes, hides when unauthenticated / on
petition / on auth routes).
- **Results match hero** (`AnalysisDisplay.tsx`) — the linked
circulating-initiative match now *leads* the analysis: a gold-accented
card with the official title, status, provenance badge (AI-matched /
user-linked), plain-language summary, and "what you'd be signing,"
instead of a dim-headed section buried below the OCR text. Additional
matches drop to compact cards below. **This is what the new
circulating-initiative scrape feeds** (e.g. 26-0005).
- **Landing rebrand** — sunflower mark, benefit-first headline ("Know
what you're signing"), one gold CTA with clear secondary actions. All
strings i18n'd (en/es).
- **Blue → gold** — processing spinner and key-point bullets move onto
the brand accent.
- **Fix: "My Scans" was broken** — its `MyScanHistory` query declared
`$skip`/`$take` as `Float!` while the field args are `Int!`, so the page
errored (`Variable "$skip" of type "Float!" used in position expecting
type "Int!"`). Pre-existing; fixed to `Int!`.

## Deferred (own follow-up — needs a rendered visual pass)
Moving the non-camera surfaces off pure black onto the paper/ink canvas
is an all-or-nothing token change across 5 screens; I want it eyeballed
rendered before shipping. The processing **stepper** is also deferred.
Camera screen stays black. This increment keeps the current dark
surface.

## Verification
`pnpm --filter frontend test` green (incl. the updated
AnalysisDisplay/results tests and the new ScanFab test); tsc clean; lint
clean; a11y labels in place. **Recommend a `pnpm test:a11y` + visual
pass** before merge. Self-review (sole author) — needs a countersign.

## Design reference
Mockup: the approved 4-screen flow (landing / capture / processing /
results hero).

+++ apps/frontend/lib/graphql/documents.ts
@@ -489,8 +489,8 @@ export interface DeleteAllScansResult {
 
 export const GET_MY_SCAN_HISTORY = gql`
   query MyScanHistory(
-    $skip: Float!
-    $take: Float!
+    $skip: Int!
+    $take: Int!
     $filters: ScanHistoryFiltersInput
   ) {
     myScanHistory(skip: $skip, take: $take, filters: $filters) {
